really minor changes

+++ app/src/main/java/com/example/udrink/Adapters/DrinkFeedAdapter.java
@@ -101,7 +101,7 @@ private ChildEventListener mChildListener;
         // - get element from your dataset at this position
         // - replace the contents of the view with that element
         holder.beerName.setText(mDataset.get(position).getDrinkName());
-        holder.beerABV.setText(Integer.toString(mDataset.get(position).getABV()));
+        holder.beerABV.setText(Double.toString(mDataset.get(position).getABV()));
         holder.beerOunces.setText(Integer.toString(mDataset.get(position).getOunces()));
         holder.timeAgo.setText(mDataset.get(position).getDrankAt().toString());
